refactor(store): flatten state/ into state.rs
Concatenate state/{state, store, trigger, codec_adapter, testing}.rs into
a single state.rs at the crate root. All five pieces form one cohesive
"versioned state persistence" concept — State payload, StateStore trait,
PersistTrigger policy, CodecStateStore bridge, and the InMemoryStateStore
testing fake.

The testing fake lives in an inline `#[cfg(feature = "testing")] mod
testing { ... }` block, matching the codec.rs `pub mod serde` pattern
used for feature-gated nested content.

Public paths unchanged:
- nexus_store::state::{State, StateStore, CodecStateStore,
  CodecStateStoreError, PersistTrigger, EveryNEvents, AfterEventTypes}
- nexus_store::state::InMemoryStateStore (feature = "testing")
