--- crates/nexus-store/src/state/mod.rs
@@ -1,13 +1,0 @@
-mod codec_adapter;
-mod state;
-mod store;
-#[cfg(feature = "testing")]
-mod testing;
-mod trigger;
-
-pub use self::state::State;
-pub use codec_adapter::{CodecStateStore, CodecStateStoreError};
-pub use store::StateStore;
-#[cfg(feature = "testing")]
-pub use testing::InMemoryStateStore;
-pub use trigger::{AfterEventTypes, EveryNEvents, PersistTrigger};
